backend/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langchain-community
langchain-chroma
chromadb
pypdf
python-dotenv
tiktoken
google-generativeai